move old circuit to a module
