+++ crates/light-client-prover/src/circuit/mod.rs
@@ -1,0 +1,2 @@
+// Until we transition to the new system completely, we will keep the old circuit
+pub mod old;
