+++ guests/risc0/light-client-proof/bitcoin/src/bin/light_client_proof_bitcoin.rs
@@ -1,7 +1,7 @@
 #![no_main]
 use bitcoin_da::spec::RollupParams;
 use bitcoin_da::verifier::BitcoinVerifier;
-use citrea_light_client_prover::circuit::run_circuit;
+use citrea_light_client_prover::circuit::old::run_circuit;
 use citrea_primitives::{TO_BATCH_PROOF_PREFIX, TO_LIGHT_CLIENT_PREFIX};
 use citrea_risc0_adapter::guest::Risc0Guest;
 use sov_rollup_interface::da::DaVerifier;
